Move Watch class into its own files and update makefile.
