+++ grive-daemon.cpp
@@ -46,15 +46,10 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include <map>
 #include <dirent.h>
 #include "daemon.h"
-
-using std::map;
-using std::string;
-using std::cout;
-using std::endl;
+#include "watch.h"
 
 /* Allow for 1024 simultaneous events */
 #define BUFF_SIZE ((sizeof(struct inotify_event)+FILENAME_MAX)*1024)
-
 #define WATCH_FLAGS (IN_ATTRIB | IN_CREATE | IN_DELETE | IN_DELETE_SELF | IN_MODIFY | IN_MOVE_SELF | IN_MOVED_FROM | IN_MOVED_TO)
 
 int initialRecursiveWd()
@@ -63,57 +58,6 @@ int initialRecursiveWd()
   wd--;
   return wd;
 }
-
-class Watch {
-    struct wd_elem {
-        int pd;
-        string name;
-        bool operator() (const wd_elem &l, const wd_elem &r) const
-            { return l.pd < r.pd ? true : l.pd == r.pd && l.name < r.name ? true : false; }
-    };
-    map<int, wd_elem> watch;
-    map<wd_elem, int, wd_elem> rwatch;
-public:
-    // Insert event information, used to create new watch, into Watch object.
-    void insert( int pd, const string &name, int wd ) {
-        wd_elem elem = {pd, name};
-        watch[wd] = elem;
-        rwatch[elem] = wd;
-    }
-    // Erase watch specified by pd (parent watch descriptor) and name from watch list.
-    // Returns full name (for display etc), and wd, which is required for inotify_rm_watch.
-    string erase( int pd, const string &name, int *wd ) {
-        wd_elem pelem = {pd, name};
-        *wd = rwatch[pelem];
-        rwatch.erase(pelem);
-        const wd_elem &elem = watch[*wd];
-        string dir = elem.name;
-        watch.erase(*wd);
-        return dir;
-    }
-    // Given a watch descriptor, return the full directory name as string. Recurses up parent WDs to assemble name, 
-    // an idea borrowed from Windows change journals.
-    string get( int wd ) {
-        const wd_elem &elem = watch[wd];
-        return elem.pd <= -1 ? elem.name : this->get(elem.pd) + "/" + elem.name;
-    }
-    // Given a parent wd and name (provided in IN_DELETE events), return the watch descriptor.
-    // Main purpose is to help remove directories from watch list.
-    int get( int pd, string name ) {
-        wd_elem elem = {pd, name};
-        return rwatch[elem];
-    }
-    void cleanup( int fd ) {
-        for (map<int, wd_elem>::iterator wi = watch.begin(); wi != watch.end(); wi++) {
-            inotify_rm_watch( fd, wi->first );
-            watch.erase(wi);
-        }
-        rwatch.clear();
-    }
-    void stats() {
-        cout << "number of watches=" << watch.size() << " & reverse watches=" << rwatch.size() << endl;
-    }
-};
 
 struct DirectoryReader
 {
